+++ studyAi/package.json
@@ -50,8 +50,8 @@
     "reflect-metadata": "0.1.13",
     "swiper": "11.0.3",
     "uuid": "9.0.1",
-    "zod": "3.22.4"
-    "react-icons": "4.11.0",
+    "zod": "3.22.4",
+    "react-icons": "4.11.0"
   },
   "devDependencies": {
     "@graphql-codegen/cli": "5.0.0",
